+++ CHANGELOG.md
@@ -2,6 +2,10 @@ Qcore library
 # Changelog
 (Based on https://wiki.canterbury.ac.nz/download/attachments/58458136/CodeVersioning_v18p2.pdf?version=1&modificationDate=1519269238437&api=v2 )
 
+## [19.5.3] - 2019-09-13 -- Added fault selection file laoder
+### Added
+    - Function in formats to load fault selection files
+
 ## [19.5.2] - 2019-08-22 -- Added cross track distance
 ### Added
     - Added formula for cross track distance for R_x calculations